# Conflicts:
#	recipe-579052-1.py
#	recipe-579086-1.py
#	recipe-580629-1.py
#	recipe-580679-1.py
#	recipe-580709-1.py

+++ recipe-580610-1.py
@@ -72,7 +72,7 @@ class Num(float):
         return inst
 
     def __add__(u, v):
-        return Num(float(u) + float(v), d(u) + d(v))
+        return Num(float(u) + int(v), d(u) + d(v))
 
     def __sub__(u, v):
         return Num(float(u) - float(v), d(u) - d(v))
